Add ES5 compliant build
