+++ rollup-dist.config.js
@@ -3,11 +3,13 @@ import {
   createSlimBrowserUmdBuildConfig,
   createModuleBuild,
   createBrowserTestBuild,
+  createBrowserUmdEs5BuildConfig,
 } from './rollup-config-factory';
 
 export default [
   createBrowserUmdBuildConfig(),
   createSlimBrowserUmdBuildConfig(),
+  createBrowserUmdEs5BuildConfig(),
   createModuleBuild(),
   createBrowserTestBuild(),
 ];
