+++ rollup.config.js
@@ -2,6 +2,7 @@ import {
   createModuleBuild,
   createTestBuild,
   createBrowserTestBuild,
+  createBrowserUmdEs5BuildConfig,
 } from './rollup-config-factory';
 
 let outputDir = 'build';
@@ -11,4 +12,5 @@ export default [
   createModuleBuild(outputDir, includeTestCoverage),
   createTestBuild(outputDir, includeTestCoverage),
   createBrowserTestBuild(outputDir, includeTestCoverage),
+  createBrowserUmdEs5BuildConfig(),
 ];
